Rename powerpoint to insertion
